requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.7.0
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
docopt==0.6.2
-e git+https://github.com/Stefan4472/portfolio-analyzer.git@fafa2d9c8c79c75956e96d4e1ea372411e1f6f8f#egg=finance_cache&subdirectory=finance-cache
Flask==2.3.2
frozendict==2.3.10
greenlet==3.0.2
html5lib==1.1
idna==3.6
isort==5.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pathspec==0.11.2
peewee==3.17.0
pep517==0.13.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pydantic==2.5.2
pydantic_core==2.14.5
pyrate-limiter==3.1.0
python-dateutil==2.8.2
python-dotenv~=1.0.0
pytz==2023.3.post1
requests==2.31.0
requirementslib==3.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9
yfinance==0.2.33
